Moved backend files into /src to further organize.
Edited some config file paths and the ./run in root to deploy server and website accordingly inside EC2.
