+++ backend/src/models/scores/scorecard.ts
