+++ backend/src/tests/metrics/licenseMetric.test.ts
