+++ backend/src/tests/server/endpoints/package.test.ts
@@ -5,7 +5,7 @@
 
 import { expect, describe, it } from "vitest";
 import request from "supertest";
-import app from "../../../server/server.js";
+import app from "../../../src/server/server.js";
 
 describe("/package endpoint", () => {
   // Missing fields in package data
