--- backend/utils/reports/extract-results.ts
@@ -1,45 +1,0 @@
-/**
- * @file extract-results.ts
- * @description This script reads the test results and coverage data from the reports directory
- * and writes the results to the console.
- */
-
-import fs from "fs";
-
-function readTestResults(): { totalTests: number; passedTests: number } {
-  try {
-    const resultsPath = "src/utils/reports/vitest-results.json";
-    const data = fs.readFileSync(resultsPath, "utf-8");
-    const results = JSON.parse(data);
-
-    const totalTests = results.numTotalTests;
-    const passedTests = results.numPassedTests;
-
-    return { totalTests, passedTests };
-  } catch {
-    return { totalTests: 0, passedTests: 0 };
-  }
-}
-
-function readCoverageData(): number {
-  try {
-    const coveragePath = "src/utils/reports/coverage/coverage-summary.json";
-    const data = fs.readFileSync(coveragePath, "utf-8");
-    const coverage = JSON.parse(data);
-
-    return coverage.total.lines.pct;
-  } catch {
-    // Return a default value if the file is not found (some tests have failed)
-    return 87.7;
-  }
-}
-
-function writeResultsToConsole() {
-  const { totalTests, passedTests } = readTestResults();
-  const coveragePercentage = readCoverageData();
-
-  const statement = `${passedTests}/${totalTests} test cases passed. ${coveragePercentage}% line coverage achieved.`;
-  console.log(statement.trim());
-}
-
-writeResultsToConsole();
