+++ client/src/App.tsx
@@ -1,4 +1,5 @@
-/* eslint-disable react/react-in-jsx-scope */
+
+
 const App = () => {
   return <div>Group 5 - Trustworthy Module Registry Front End Holder</div>;
 };
