+++ run
@@ -26,10 +26,16 @@ fi
 
 # Check if the argument is a file path
 if [ -f "$arg" ]; then
-  # Compile TS files
+  # Compile TS files && Vite build
+  cd client
   npm run build
-  # Pass the file path index.js
-  node dist/server/server.js --file "$arg"
+  # Copy the "build" folder to "/var/www/html" for Nginx
+  sudo cp -R build/ /var/www/html/front-end/
+  # Compile TS files
+  cd ../backend
+  npm run build
+  # Start Server
+  npm run start
   exit 0
 fi
